Xvnc: support ExtendedMouseButtons pseudo-encoding
This commit adds support for the pseudo-encoding ExtendedMouseButtons in
Xvnc, which makes it possible to use to use the back/forward mouse
buttons.

This commit contains work originally done by
PixelSmith <manny33@frontbuffer.com>.

+++ common/rfb/VNCSConnectionST.cxx
@@ -477,7 +477,7 @@ void VNCSConnectionST::setPixelFormat(const PixelFormat& pf)
   setCursor();
 }
 
-void VNCSConnectionST::pointerEvent(const Point& pos, uint8_t buttonMask)
+void VNCSConnectionST::pointerEvent(const Point& pos, uint16_t buttonMask)
 {
   if (rfb::Server::idleTimeout)
     idleTimer.start(secsToMillis(rfb::Server::idleTimeout));
